+++ orte/libjorte/JORTETypeRegisterDestroyAll.c
@@ -28,13 +28,14 @@
 
 // library header file's path
 #include "orte.h"
+#include "jorte/4all.h"
 // pregenerated header
 #include "jorte/org_ocera_orte_DomainApp.h"
 
 
 JNIEXPORT jboolean JNICALL
 Java_org_ocera_orte_DomainApp_jORTETypeRegisterDestroyAll
-(JNIEnv *env, jclass cls, jint j_app_domain_handle)
+(JNIEnv *env, jclass cls, jlong j_app_domain_handle)
 {
   int b;
